# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/user/token": {
      "get": {
        "summary": "生成访问令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "用户管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/token/": {
      "get": {
        "summary": "获取所有令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "post": {
        "summary": "创建令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "put": {
        "summary": "更新令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/token/{id}": {
      "get": {
        "summary": "获取指定令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "delete": {
        "summary": "删除令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/usage/token/": {
      "get": {
        "summary": "获取令牌使用情况",
        "deprecated": false,
        "description": "🔑 需要令牌认证（TokenAuth）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/log/token": {
      "get": {
        "summary": "通过令牌获取日志",
        "deprecated": false,
        "description": "🔓 无需鉴权（通过令牌查询）",
        "tags": [
          "日志"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Token": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "expired_time": {
            "type": "integer"
          },
          "remain_quota": {
            "type": "integer"
          },
          "unlimited_quota": {
            "type": "boolean"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken1": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination222": {
        "group": [
          {
            "id": 573666
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
          {
            "id": 573667
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}